+++ http-client-tls/ChangeLog.md
@@ -1,3 +1,7 @@
+## 0.2.4
+
+* Global manager
+
 ## 0.2.3
 
 * Exception catching cleanup